refactor: move authenticate browser func to utils
